+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Chevre決済中止処理のエラーハンドリングを調整
 - 予約使用レスポンスにアクションIDを含めることができるように調整
 - 予約に対する使用アクション検索をChevreで実装
+- 予約使用時にagent.identifierを指定できるように調整
 
 ### Deprecated